Consolidated UnitSystem into GizmoObjectManipulation.hpp/.cpp
Following advising from Collaborator yw4z, I consolidated the contents of the UnitConversion singleton away from their own file and into GizmoObjectManipulation.hpp/cpp.

This change involves removing the separate files of UnitCoversion.hpp/cpp, and adjusting the CMakeLists.txt accordingly, and removing #includes of UnitCoversion.hpp for a cleaner submission.

The singleton is necessary beyond the existing app_config pattern because the Alt-hold temporary swap requires transient state that shouldn't be written to disk.  It also includes to broadcast changes to listeners for decoupled communication.
