+++ src/slic3r/CMakeLists.txt
@@ -454,8 +454,6 @@ set(SLIC3R_GUI_SOURCES
     GUI/TextLines.hpp
     GUI/TickCode.cpp
     GUI/TickCode.hpp
-    GUI/UnitConversion.cpp
-    GUI/UnitConversion.hpp
     GUI/UnsavedChangesDialog.cpp
     GUI/UnsavedChangesDialog.hpp
     GUI/UpdateDialogs.cpp
